Physics + new Mesh system + docking

+++ GDENG03-Activities/Code/WinMain.cpp
@@ -1,13 +1,14 @@
 #include "GameEngine/GameEngineWindow.h"
 #include <time.h>
 
+
 int CALLBACK WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine, int nCmdShow)
 {
-	srand(time(NULL));
+	srand((unsigned int)time(NULL)); 
 
 	GameEngineWindow gameEngine(60);  
 
-	if (gameEngine.Init(960, 720, "Aamir Akim"))  
+	if (gameEngine.Init(960, 720, "Aamir Akim")) 
 	{
 		while (gameEngine.IsRunning())  
 		{
@@ -15,7 +16,7 @@ int CALLBACK WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLi
 		}
 	}
 
-	gameEngine.Release();  
+	gameEngine.Release();
 
 	return 0;
 }
